refactor: 约定了folder命名规则
