+++ .umirc.ts
@@ -27,11 +27,11 @@ export default defineConfig({
       },
       {
         title: '组合组件',
-        children: getMds('components/display').map((item) => 'components/display/' + item),
+        children: getMds('components/composit').map((item) => 'components/composit/' + item),
       },
       {
         title: '业务套件',
-        children: getMds('components/composit').map((item) => 'components/composit/' + item),
+        children: getMds('components/business/').map((item) => 'components/business/' + item),
       },
     ],
   },
