+++ docs/components/composit/detail-card.md
