+++ docs/demo/menu/head-menu.demo.tsx
@@ -3,8 +3,8 @@
  * desc: 水平的顶部导航菜单。
  */
 import React from 'react';
-import Menu from '@/Menu/Menu';
-import Icon from '@/Icon/Icon';
+import Menu from '@/menu';
+import Icon from '@/icon';
 
 const { SubMenu } = Menu;
 class App extends React.PureComponent {
