+++ docs/demo/menu/side-menu.demo.tsx
@@ -3,9 +3,9 @@
  * desc: 垂直菜单，子菜单内嵌在菜单区域。
  */
 import React from 'react';
-import Menu from '@/Menu/Menu';
-import Icon from '@/Icon/Icon';
-import Button from '@/Button/Button';
+import Menu from '@/menu';
+import Icon from '@/icon';
+import Button from '@/button';
 
 const { SubMenu } = Menu;
 class App extends React.PureComponent {
