+++ docs/demo/table/with-search.demo.tsx
@@ -3,8 +3,8 @@
  * desc: 通过传入 `data` 来展示表格，通过 `onChange`和 `onSearch` 等参数控制交互。
  */
 import React from 'react';
-import Table from '@/Table/Table';
-import { Slider } from 'antd';
+import Table from '@/table';
+import Slider from '@/slider';
 import axios from 'axios';
 
 const columns = [
