+++ docs/demo/table/with-toolbar.demo.tsx
@@ -3,9 +3,9 @@
  * desc: 通过传入 `tools` 组件数组，来展示对应的交互工具。
  */
 import React from 'react';
-import Table from '@/Table/Table';
-import Button from '@/Button/Button';
-import Icon from '@/Icon/Icon';
+import Table from '@/table';
+import Button from '@/button';
+import Icon from '@/icon';
 import axios from 'axios';
 
 const columns = [
