+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@mlz/admin",
-  "version": "0.3.1",
+  "version": "0.4.12",
   "description": "一套编程猫设计规范下的管理系统React组件库，基于Antd",
   "license": "MIT",
   "keywords": [
