+++ src/Avatar/index.tsx
