+++ src/Slider/index.tsx
