+++ src/Table/__tests__/Table.test.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { mount, shallow } from 'enzyme';
-import Table from '../Table';
+import Table from '..';
 
 const createColumns = ($type = null) => {
   return [
