+++ src/detail-card/index.less
