+++ src/detail-card/index.type.ts
@@ -1,8 +1,8 @@
 // <reference types="react" />
 import React from 'react';
 import { CardProps } from 'antd/lib/card';
-import { IColumnTypes } from '../Table/common-table/index.type';
-export { SupporttedColumnTypes } from '@/Table/common-table/index.type';
+import { IColumnTypes } from '../table/common-table/index.type';
+export { SupporttedColumnTypes } from '@/table/common-table/index.type';
 import { TableProps as RcTableProps } from 'rc-table/lib/Table';
 
 export interface IDetailCardColumn {
