+++ src/index.tsx
@@ -1,27 +1,27 @@
-// export * from 'antd';
-export { default as Button } from './Button/Button';
-export { default as Avatar } from './Avatar/Avatar';
-export { default as Card } from './Card/Card';
-export { default as Col } from './Col/Col';
-export { default as ConfigProvider } from './ConfigProvider/ConfigProvider';
-export { default as DatePicker } from './DatePicker/DatePicker';
-export { default as Form } from './Form/Form';
-export { default as Icon } from './Icon/Icon';
-export { default as Input } from './Input/Input';
-export { default as InputNumber } from './InputNumber/InputNumber';
-export { default as Layout } from './Layout/Layout';
-export { default as Menu } from './Menu/Menu';
-export { default as Modal } from './Modal/Modal';
-export { default as PageHeader } from './PageHeader/PageHeader';
-export { default as Popconfirm } from './Popconfirm/Popconfirm';
-export { default as Popover } from './Popover/Popover';
-export { default as Radio } from './Radio/Radio';
-export { default as Result } from './Result/Result';
-export { default as Select } from './Select/Select';
-export { default as Slider } from './Slider/Slider';
-export { default as Spin } from './Spin/Spin';
-export { default as Table } from './Table/Table';
-export { default as Tag } from './Tag/Tag';
-export { default as Tooltip } from './Tooltip/Tooltip';
-export { default as Upload } from './Upload/Upload';
-export { default as DetailCard } from './DetailCard/DetailCard';
+export * from 'antd';
+export { default as Button } from './button';
+export { default as Avatar } from './avatar';
+export { default as Card } from './card';
+export { default as Col } from './col';
+export { default as ConfigProvider } from './config-provider';
+export { default as DatePicker } from './date-picker';
+export { default as Form } from './form';
+export { default as Icon } from './icon';
+export { default as Input } from './input';
+export { default as InputNumber } from './input-number';
+export { default as Layout } from './layout';
+export { default as Menu } from './menu';
+export { default as Modal } from './modal';
+export { default as PageHeader } from './page-header';
+export { default as Popconfirm } from './popconfirm';
+export { default as Popover } from './popover';
+export { default as Radio } from './radio/radio';
+export { default as Result } from './result';
+export { default as Select } from './select';
+export { default as Slider } from './slider';
+export { default as Spin } from './spin';
+export { default as Table } from './table';
+export { default as Tag } from './tag';
+export { default as Tooltip } from './tooltip';
+export { default as Upload } from './upload';
+export { default as DetailCard } from './detail-card';
